% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slides.R
\name{get_slides}
\alias{get_slides}
\title{Load the background slides}
\usage{
get_slides()
}
\description{
This is a helper function that simply tries
to load the background slides for the experiment.
In a local environment, this function will likely
load the slides into Powerpoint. In others, it may
just "download" the slides which you can open yourself.
}
\details{
This is an example of how you can include arbitrary files
in R packages.
}
\examples{
\dontrun{
get_slides()
}

}
\references{
\url{http://starbuck1.s3.amazonaws.com/expression-prediction/Expression_prediction_slides.pptx}
}
